+++ packer-build.ps1
@@ -1,0 +1,2 @@
+# run packer for vagrant virtualbox provider
+packer build packer-virtualbox.json