+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
-## 1.0.0
+## 1.0.x
+
 **Implemented enhancements:**
-- Initial version
+
+- k3s installation
+
+## 2.0.x
+
+**Implemented enhancements:**
+
+- Packer build implemented